Updates for 0.2.1
Тэги, фильтры в поиске, поломка в чаше, юзернейм вспывашка, слайдер большой, отызвы фикс, контакты  скрываются, алерт месседжы для по локации и картинкам

DT-401 DT-406 DT-403 DT-389

+++ Dagestan/Targets/DagestanTrails/Sources/Data/Places/PlaceDetailDTO.swift
@@ -47,7 +47,7 @@ extension PlaceDetailDTO {
         let title: String
         let shortDescription: String?
         let images: [ImageDTO]
-        let rating: Int
+        let rating: Double
         let distance: Double
         let travelTime: String
         let isFavorite: Bool
